digital-trace/digital-trace/requirements.txt
flask
requests
python-whois
Pillow
python-docx
colorama
tabulate
jinja2